+++ test/dummy_app/config/routes.rb
@@ -1,4 +1,9 @@
 Dummy::Application.routes.draw do
+  get "reports/custom"
+  get "reports/custom_disposition"
+  get "reports/custom_filename"
+  get "reports/custom_headers"
+  get "reports/ivar_filename"
   get "reports/sample"
   get "reports/table"
 end