+++ test/prawn_rails_test.rb
@@ -7,8 +7,6 @@ class PrawnRailsTest < ActiveSupport::TestCase
   include PrawnRails::RailsHelper
 
   test "text can take non strings" do
-    @filename = ""
-
     output = prawn_document do |pdf|
       pdf.text "Number:"
       pdf.text 10
@@ -19,5 +17,12 @@ class PrawnRailsTest < ActiveSupport::TestCase
 
     page = reader.page(1).to_s
     assert page.include?("Number:\n10")
+  end
+
+  test "matches .PDF extension regardless of case" do
+    ["pDf", "pdf", "PDF"].each do |ext|
+      @filename = "test.#{ext}"
+      assert_equal get_metadata[:info][:Title], "test"
+    end
   end
 end